feat: website-file-breakdown-for-current-selection-001 - Establish current-selection/base shared package
Introduce base/ for shared selection contracts, cross-kind messages, layout/locale chrome, and no-selection UI with a public barrel and legacy-path re-export shims so existing imports keep working during the package split.
