+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -93,6 +93,7 @@ work inputs.
 - When graph-editor draft validation needs duplicate identifier detection, count node IDs from the pending editable-definition entity arrays before converting them into keyed maps; `Map`-indexed topology builders collapse duplicates and cannot prove pre-save duplicate-name errors on their own.
 - When the dashboard needs selected-workstation edit values, resolve the selected runtime node against the canonical `factory.workstations` list by `transition_id` or workstation name, then join that workstation to `factory.workers` through `workstation.worker`; prompt lives on workstation `body`, template on `promptFile`, and model on the referenced worker.
 - When the dashboard opens a workstation edit session, hydrate the draft from canonical editable values once per selected workstation and only rehydrate on later current-factory refreshes if the local draft is still clean; dirty drafts should keep the customer's in-progress values until an explicit save or reset path handles the conflict.
+- Shared current-selection selection contracts, history store, resolution helpers, cross-kind message catalogs, detail layout/locale chrome, and `NoSelectionDetailCard` live under `ui/src/features/current-selection/base/` and export through `ui/src/features/current-selection/base/public`; legacy deep paths remain as thin re-export shims until cross-feature imports migrate in the package-breakdown follow-up stories.
 - When the dashboard saves current factory edits, use the canonical `PUT /factory-sessions/{session_id}/factory` replacement route and keep the typed current-factory adapter local to `ui/src/api/current-factory-definition/api.ts`; keep mutation orchestration in `ui/src/features/current-selection/` and update both seams on success: write the normalized saved `Factory` document into the `current-editable-factory-definition` React Query cache and mark the local workstation draft clean in the same success path so the header save action, form status, and saved values stay aligned without a second local shadow copy.
 - When editable-definition saves reject topology before persistence, return `apisurface.TopologyValidationError` targets that map graph issues to `node`, `edge`, `field`, or `save` locations; the API handler should preserve those targets in the `INVALID_FACTORY` response instead of collapsing every validation failure to the form root.
 - When the current-selection workstation editor narrows to workstation-scoped fields such as `worker` and `prompt`, keep the mutation path anchored on the canonical current-factory definition but only apply the fields that truly belong to the selected workstation; preserve unrelated worker-model and template data from the latest fetched definition instead of carrying stale hidden values forward into the editable-definition save payload.
