+++ ui/src/features/current-selection/base/components/current-selection-detail-layout.test.tsx
