+++ ui/src/features/current-selection/base/components/current-selection-locale.test.tsx
