+++ ui/src/features/current-selection/base/components/no-selection-detail-card.test.tsx
