+++ ui/src/features/current-selection/base/messages/current-selection-dispatch-history.test.ts
@@ -1,4 +1,4 @@
-import { SUPPORTED_LOCALES } from "../../../i18n";
+import { SUPPORTED_LOCALES } from "../../../../i18n";
 import {
   currentSelectionDispatchHistoryMessagesByLocale,
   getCurrentSelectionDispatchHistoryMessages,
