+++ ui/src/features/current-selection/base/messages/current-selection-shell.test.ts
@@ -1,4 +1,4 @@
-import { SUPPORTED_LOCALES } from "../../../i18n";
+import { SUPPORTED_LOCALES } from "../../../../i18n";
 import {
   currentSelectionShellMessagesByLocale,
   getCurrentSelectionShellMessages,
