+++ ui/src/features/current-selection/base/state/selection-types.ts
@@ -1,0 +1,15 @@
+import type {
+  DashboardWorkItemRef,
+} from "../../../../api/dashboard/types";
+export type { TerminalWorkDetail } from "../../../terminal-work/lib/types";
+
+export type {
+  DashboardSelection,
+  DashboardWorkItemSelection,
+  DashboardWorkstationRequestSelection,
+} from "./dashboardSelection";
+
+export interface StatePositionWorkItem extends DashboardWorkItemRef {
+  startedAt?: string;
+  started_at?: string;
+}
