+++ ui/src/features/current-selection/base/state/selectionHistoryStore.test.ts
